backend/main/requirements.txt
fastapi>=0.115.13
uvicorn[standard]>=0.34.3
jinja2>=3.1.2
python-multipart>=0.0.20
starlette>=0.27.0
prometheus-client>=0.19.0
structlog>=23.2.0
sqlalchemy>=2.0.41
bcrypt>=4.3.0
python-jose[cryptography]>=3.4.0
email-validator>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
pydantic>=2.11.7
slowapi>=0.1.9
passlib>=1.7.4
httpx>=0.28.1
PyJWT>=2.10.1
watchdog>=3.0.0
openai>=1.98.0
sentence-transformers>=5.0.0
numpy>=1.21.0
python-telegram-bot>=22.3
aiohttp>=3.12.15
python-dotenv>=1.0.0
